x64 Imgui

+++ Core/GpuBuffer.cpp
@@ -32,7 +32,7 @@ namespace YiaEngine:: Graphic
             
 	}
 
-    D3D12_VERTEX_BUFFER_VIEW GpuBuffer::VertexBufferView(size_t offset, size_t stride, size_t dataSize)
+    D3D12_VERTEX_BUFFER_VIEW GpuBuffer::VertexBufferView(size_t offset, UINT stride, UINT dataSize)
     {
         D3D12_VERTEX_BUFFER_VIEW vbo;
         vbo.BufferLocation = gpuVirtualAddress_;
@@ -40,7 +40,7 @@ namespace YiaEngine:: Graphic
         vbo.SizeInBytes = dataSize;
         return vbo;
     }
-    D3D12_INDEX_BUFFER_VIEW GpuBuffer::IndexBufferView(size_t offset, size_t stride, size_t dataSize)
+    D3D12_INDEX_BUFFER_VIEW GpuBuffer::IndexBufferView(size_t offset, UINT stride, UINT dataSize)
     {
         D3D12_INDEX_BUFFER_VIEW ibo;
         ibo.BufferLocation = gpuVirtualAddress_ + offset;
